# models/model.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Model(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: str=None, name: str=None, description: str=None, file_url: str=None):  # noqa: E501
        """Model - a model defined in Swagger

        :param id: The id of this Model.  # noqa: E501
        :type id: str
        :param name: The name of this Model.  # noqa: E501
        :type name: str
        :param description: The description of this Model.  # noqa: E501
        :type description: str
        :param file_url: The file_url of this Model.  # noqa: E501
        :type file_url: str
        """
        self.swagger_types = {
            'id': str,
            'name': str,
            'description': str,
            'file_url': str
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'description': 'description',
            'file_url': 'file_url'
        }
        self._id = id
        self._name = name
        self._description = description
        self._file_url = file_url

    @classmethod
    def from_dict(cls, dikt) -> 'Model':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Model of this Model.  # noqa: E501
        :rtype: Model
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this Model.

        Уникальный идентификатор модели  # noqa: E501

        :return: The id of this Model.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this Model.

        Уникальный идентификатор модели  # noqa: E501

        :param id: The id of this Model.
        :type id: str
        """

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this Model.

        Название модели  # noqa: E501

        :return: The name of this Model.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Model.

        Название модели  # noqa: E501

        :param name: The name of this Model.
        :type name: str
        """

        self._name = name

    @property
    def description(self) -> str:
        """Gets the description of this Model.

        Описание модели  # noqa: E501

        :return: The description of this Model.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description: str):
        """Sets the description of this Model.

        Описание модели  # noqa: E501

        :param description: The description of this Model.
        :type description: str
        """

        self._description = description

    @property
    def file_url(self) -> str:
        """Gets the file_url of this Model.

        URL файла модели  # noqa: E501

        :return: The file_url of this Model.
        :rtype: str
        """
        return self._file_url

    @file_url.setter
    def file_url(self, file_url: str):
        """Sets the file_url of this Model.

        URL файла модели  # noqa: E501

        :param file_url: The file_url of this Model.
        :type file_url: str
        """

        self._file_url = file_url